fix: attempt to fix garbage output
